pyproject.toml
[tool.poetry]
name = "orto-tezzi"
version = "0.1.0"
description = ""
authors = ["Ahmad Umaev <ahmad.um.2001@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

asgiref = "^3.6.0"
attrs = "^22.2.0"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.1"
inflection = "^0.5.1"
jsonschema = "^4.17.3"
Pillow = "^9.5.0"
psycopg2-binary = "^2.9.5"
pyrsistent = "^0.19.3"
pytz = "^2023.3"
PyYAML = "^6.0"
sqlparse = "^0.4.3"
uritemplate = "^4.1.1"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
